Refactored _util to _utils package with log for logging
